Added Multi-Base Spawning for Server
Now, when the host/server joins the game, multiple bases will spawn as enemy bases to all players. Bases spawn randomly and at a minimum distance from all other bases (to prevent them from spawning closely or on top of each other.)

Additionally, a second unit was added. By doing so, noticeable problems in movement and agent pathing have come up due to the unit's larger size. That is something that needs to be addressed and possibly rewritten in the near future.