Upgraded Minimap
The minimap has been redone from scratch. Originally, a top down camera was used on the entire map as the minimap. However, this introduced visual problems, as the camera was trying to render large and far-away objects on a small rendered image.

Now, that system has been replaced with a small replica of the map, with the top down camera focused on the replica instead of the whole map. This way, map icons are always visible with 100% sharpness, with no anti-aliasing required.